Update typedoc to fix handlebars 4.6.x breakage
As per https://github.com/TypeStrong/typedoc/issues/1159, in response to
security concerns, handlebars 4.6.0 changed in a way that broke our
version of typedoc. This was fixed and made available in 0.15.7.

This is fortunate as typedoc 0.16.0 makes a breaking change that exposes
a lot more than I think we want to:
https://github.com/TypeStrong/typedoc/releases/tag/v0.16.0

The current lead maintainer outlines a new vision for a "library" mode
here which may be of interest in future:
https://github.com/TypeStrong/typedoc/issues/1266

Signed-off-by: Thomas Chetwin <tchetwin@bloomberg.net>